+++ src/workforce/connectors/index.ts
@@ -1,0 +1,31 @@
+export { WORKFORCE_BRAIN_CONNECTIONS } from './connectionCatalog';
+export {
+  getWorkforceFinanceContribution,
+  publishWorkforceFinanceContribution,
+  type WorkforceFinanceContribution,
+} from './financeBridge';
+export {
+  getAppointmentBrainRoster,
+  publishAppointmentRosterRefresh,
+  listBookableEmployees,
+  employeeAvailableAt,
+  getWorkingHoursForDay,
+  type AppointmentBrainRosterSnapshot,
+  type BookableEmployee,
+} from './appointmentBridge';
+export {
+  getCustomerServiceCapacity,
+  publishCustomerCapacitySignal,
+  type CustomerBrainServiceCapacity,
+} from './customerBridge';
+export {
+  getWorkforceHealthSelfCheck,
+  publishWorkforceHealthSelfCheck,
+  type WorkforceHealthSelfCheck,
+} from './healthBridge';
+export {
+  getOwnerWorkforceDashboard,
+  type OwnerWorkforceDashboard,
+  type OwnerWorkforceAttentionItem,
+} from './ownerBridge';
+export { publishWorkforceEcosystemRefresh } from './ecosystemRefresh';